Fixed spring-mvc-java, spring-cloud-rest-discovery-server apache-solrj integration tests
