+++ spring-mvc-java/src/main/java/com/baeldung/web/controller/WebsocketSendToUserController.java
@@ -1,22 +1,18 @@
 package com.baeldung.web.controller;
 
-import com.google.gson.Gson;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.handler.annotation.MessageExceptionHandler;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.messaging.simp.SimpMessageSendingOperations;
-import org.springframework.messaging.simp.annotation.SendToUser;
-import org.springframework.stereotype.Controller;
-
 import java.security.Principal;
 import java.util.Map;
 
+import org.springframework.messaging.handler.annotation.MessageExceptionHandler;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.handler.annotation.Payload;
+import org.springframework.messaging.simp.annotation.SendToUser;
+import org.springframework.stereotype.Controller;
+
+import com.google.gson.Gson;
+
 @Controller
 public class WebsocketSendToUserController {
-
-    @Autowired
-    private SimpMessageSendingOperations messagingTemplate;
 
     private Gson gson = new Gson();
